forecastingHorizon moved to from Forecaster to metric
